requirements.txt
selenium
beautifulsoup4
pandas
pyyaml
requests
pillow
